+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Memento
-### [2.3.2](https://github.com/diomsg-code/Memento/tree/2.3.2) (2024-12-17)
-[Full Changelog](https://github.com/diomsg-code/Memento/compare/2.3.1...2.3.2) - [Previous Releases](https://github.com/diomsg-code/Memento/releases)
+### [2.3.3](https://github.com/diomsg-code/Memento/tree/2.3.3) (2024-12-26)
+[Full Changelog](https://github.com/diomsg-code/Memento/compare/2.3.2...2.3.3) - [Previous Releases](https://github.com/diomsg-code/Memento/releases)
 
-- Updated: TOC version for patch 11.0.7 (Retail)
+- Changed: For the 'Battleground completed' event, you can now decide whether the screenshots should only be taken in the event of a victory
+- Changed: For the 'Brawl completed' event, you can now decide whether the screenshots should only be taken in the event of a victory